requirements.txt
pykrx
pandas
fpdf2
matplotlib
